tests revised part V, (1,2,4,5,6,7 work)
